+++ src/components/DataDisplay/Chip/stories/Chip.tsx
@@ -1,5 +1,0 @@
-import { Chip as Component, ChipProps } from "@mui/material";
-
-export const Chip = (props: ChipProps) => <Component {...props} />;
-
-export default Chip;